memoise reward background carousel in parent
