+++ web/fluidity.money/src/screens/Reward/Reward.module.scss
@@ -52,3 +52,46 @@
   transition: opacity 1s ease;
   z-index: 0;
 }
+
+.winner {
+  display: flex;
+  flex-direction: row;
+  cursor: pointer;
+  font-size: 16px;
+  color: $textGrey;
+  position: relative;
+  min-width: fit-content;
+  padding: 0 10px 0 10px;
+  gap: 5px;
+  align-items: center;
+
+  a {
+    display: flex;
+    flex-direction: row;
+    cursor: pointer;
+    align-items: center;
+    gap: 5px;
+    &:hover {
+      color: $fluidWhite;
+      opacity: 0.8;
+    }
+  }
+
+  img {
+    width: 18px;
+  }
+  &:hover {
+    color: $fluidWhite;
+  }
+}
+
+.hover {
+  // cursor: pointer;
+  // &:hover {
+  //   color: white;
+  // }
+
+  &Underline {
+    text-decoration: underline;
+  }
+}
